Extract unified belief memory
