+++ v2/orchestrator-rust/src/residents/mod.rs
@@ -1,4 +1,6 @@
 mod continuity;
+mod memory;
 mod needs;
 
 pub(super) use continuity::*;
+pub(super) use memory::*;
